+++ PreOrclBackEnd/PreOrclApi/DTO/UsuariosDTO.cs
@@ -7,7 +7,7 @@ namespace PreOrclApi.DTO
 {
     public class UsuariosDTO
     {
-         public int IdUsuario { get; set; } 
+         public decimal IdUsuario { get; set; } 
         public string Usuario { get; set; }
         public string Nombre { get; set; }
         public string Apellido { get; set; }
